+++ package.json
@@ -26,6 +26,7 @@
     "sass": "^1.79.4"
   },
   "devDependencies": {
+    "@types/faker": "^6.6.9",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
